PCI: break when finding the first smaller resource
In commit d74b9027a4da (PCI: Consider additional PF's IOV BAR alignment in
sizing and assigning), it stores additional alignment in realloc_head and
take this into consideration for assignment.

After getting the additional alignment, it will re-order the head list to
make sure resources with bigger alignment is ahead of the resources with
smaller assignment in the head list. To make it happen, it iterate on the
head list and find a smaller alignment resource and insert ahead of it.
This should be done for the first occurrence, while the code now will
iterate on the whole list.

This patch fixes this behavior by break when finding the first smaller
resource in the head list.

Signed-off-by: Wei Yang <weiyang@linux.vnet.ibm.com>
Signed-off-by: Paul Mackerras <paulus@samba.org>