KVM: PPC: Book3S HV: Don't subtract timebase offset if it was not added
The dynamic micro-threading patches introduce the possibility that
by the time thread 0 of the primary vcore gets to the code that switches
to the guest, another vcore has already reached the point of asking
for an exit.  In this case, thread 0 will skip to secondary_too_late
without switching the MMU to the guest or adding the timebase offset
to the timebase.  It then goes to kvmhv_switch_to_host where it then
loads up the host MMU context and subtracts the timebase offset from
the timebase.

The result is that we return to the host context with the timebase
offset subtracted from the timebase.  If the timebase offset is
non-zero, the timebase on this core is now out of sync with the other
cores.  If the timebase offset is positive this commonly results in
a soft lockup in ktime_get_ts64() due to a negative delta from the
timebase being interpreted as a very large positive value, causing
the loop in timespec64_add_ns to take minutes to execute.  If the
timebase offset is negative then the symptoms include other random
soft lockups in both the guest and host.

The solution is simply not to do the switch-to-host code in the case
where we didn't do the switch-to-guest code.

BZ: 122589
BZ: 122590
BZ: 122872
BZ: 125326
BZ: 125662
BZ: 126320

Signed-off-by: Paul Mackerras <paulus@samba.org>