KVM: PPC: Dereference IOMMU group when releasing group
In-kernel VFIO DMA acceleration code holds the references to attached
IOMMU groups in order to make sure that the actual iommu_table's
will not go away. These references is only dropped when LIOBN is destroyed.

However in the case of VFIO-PCI guest hot-unplug, we cannot destroy
LIOBN as there may be other devices from other IOMMU groups attached
to this specific LIOBN.

This drops the additional reference when the userspace releases an IOMMU
in KVM.

Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>