book3s_hv: Disable Interrupts while updating HID0.
Disable interrupts across the update to HID0 and subsequent status
check fragment while performing dynamic split/unsplit of core. This
makes it consistent with the static split core code in which the
stop_machine frame work disables the interrupts on the CPU performing
the HID0 update.

Fixes the soft lockup in BZ 127433.

Signed-off-by: Gautham R. Shenoy <ego@linux.vnet.ibm.com>
Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>