KVM: PPC: Book3S HV: Split irq_map into all_irqs and mapped_irqs
Split the IRQ map to an irq_all_ that includes all IRQs that
are passed through and an irq_map that only contains IRQS we
want kvmppc_read_intr, the KVM real time external interrupt
exception routine to redirect.

There is a need to split this into two separate arrays because
with SRIOV and MSIx adapters, many interrupts can be passed
through, but we might want to only map a select few. Having
a large array to walk through to map the hardware IRQ to the
guest GSI (virtual IRQ) will slow interrupt handling.

The kvmppc_map_passthru_irq() routine allows a caller to
map a guest GSI - that is copy it from the irq_all to the
irq_map array. Once mapped, it is only removed when the
kvm_irq_clr_passthru_hv is called to remove the IRQ from the
list of passthrough IRQs.

Signed-off-by: Suresh Warrier <warrier@linux.vnet.ibm.com>
Signed-off-by: Paul Mackerras <paulus@samba.org>