KVM: PPC: Book3S HV: Access irq_map locklessly from KVM real mode
This patch removes the arch_spin_lock() taken in the real
mode KVM handler code when reading the map entries. This
is possible becaue each entry is only 64-bits long and the
entry if it exists is valid. Locks can be avoided by doing all
reads and stores as 64-bit operations.

Signed-off-by: Suresh Warrier <warrier@linux.vnet.ibm.com>
Signed-off-by: Paul Mackerras <paulus@samba.org>