powerpc/eeh: Set EEH_PE_RESET in eeh_pe_reset()
EEH_PE_RESET is used to indicates the state that a PE is under reset. When
PE is on host, the flag is set properly, while for a PE passed to guest,
the flag is not set during the reset.

This flag is important to VF PE, since the VF reset relies on FLR which is
issued through pci config space. VF needs to identify itself whether is in
reset procedure, so that the EEH functions won't block the config space
read/write.

This patch set/clear the EEH_PE_RESET before/after reset procedure, when a
PE is passed to guest. Otherwise, we would see "Pending transaction while
issuing FLR" during VF EEH recovery.

Signed-off-by: Wei Yang <weiyang@linux.vnet.ibm.com>
Reviewed-by: Gavin Shan <gwshan@linux.vnet.ibm.com>
Signed-off-by: Paul Mackerras <paulus@samba.org>