ipmi: Make the message handler easier to use for SMI interfaces
The message handler expected the SMI interface to keep a queue of
messages, but that was kind of silly, the queue would be easier to
manage in the message handler itself.  As part of that, fix the
message cleanup to make sure no messages are outstanding when an
SMI interface is unregistered.  This makes it easier for an SMI
interface to unregister, it just has to call ipmi_unregister_smi()
first and all processing from the message handler will be cleaned
up.

Backported commit from 7ea0ed2b5be81781ba976bc03414ef5da76270b9

Signed-off-by: Corey Minyard <cminyard@mvista.com>
Signed-off-by: Neelesh Gupta <neelegup@linux.vnet.ibm.com>
Signed-off-by: Paul Mackerras <paulus@samba.org>